add and refactor p value estimation
